+++ src/main/java/com/buuz135/simpleclaims/Main.java
@@ -110,6 +110,7 @@ public class Main extends JavaPlugin {
         interaction.register("PlaceFluid", ClaimPlaceBucketInteraction.class, ClaimPlaceBucketInteraction.CUSTOM_CODEC);
         interaction.register("RefillContainer", ClaimPickupBucketInteraction.class, ClaimPickupBucketInteraction.CUSTOM_CODEC);
         interaction.register("Replace", ClaimReplaceInteraction.class, ClaimReplaceInteraction.CUSTOM_CODEC);
+        interaction.register("ChangeBlock", ClaimChangeBlockInteraction.class, ClaimChangeBlockInteraction.CUSTOM_CODEC);
 
         partyInactivityTickingSystem = new PartyInactivityThread();
         partyInactivityTickingSystem.start();